requirements.txt
Authlib==1.4.1
blinker==1.9.0
Bootstrap-Flask==2.4.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
decorator==5.1.1
dnspython==2.7.0
Flask==3.1.0
flask-htmx==0.4.0
flask-oidc==2.2.2
Flask-WTF==1.2.2
gssapi==1.9.0
html2text==2024.2.26
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pycparser==2.22
python-freeipa==1.0.9
PyYAML==6.0.2
requests==2.32.3
requests-gssapi==1.3.0
srvlookup==3.0.0
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
